add: 实现内置限流算法可配置

+++ force-boot-starter/force-mvc-starter/src/main/java/com/quan/boot/mvc/limit/local/CounterRateLimiter.java
@@ -31,10 +31,10 @@ public class CounterRateLimiter implements RateLimiter {
     /**
      * 阈值
      */
-    private final long capacity;
+    private final int capacity;
 
 
-    public CounterRateLimiter(long time, long capacity) {
+    public CounterRateLimiter(long time, int capacity) {
         this.time = time;
         this.capacity = capacity;
     }
